third_party/WebKit/LayoutTests/inspector/elements/styles-4/styles-keyframes.html

<html>
<head>

<link rel="stylesheet" href="../styles/resources/keyframes.css">

<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<style>
@keyframes animName {
    from, 20% {
        margin-left: 200px;
        color: red;
    }
    to {
        margin-left: 500px;
    }
}
</style>
<script>
function test() {
  ElementsTestRunner.selectNodeAndWaitForStyles('element', step1);

  function step1() {
    TestRunner.addResult('=== Before key modification ===');
    ElementsTestRunner.dumpSelectedElementStyles(true);
    var section = UI.panels.elements._stylesWidget._sectionBlocks[1].sections[1];
    section.startEditingSelector();
    section._selectorElement.textContent = '1%';
    section._selectorElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
    ElementsTestRunner.waitForSelectorCommitted(step2);
  }

  function step2() {
    TestRunner.addResult('=== After key modification ===');
    ElementsTestRunner.dumpSelectedElementStyles(true);
    TestRunner.domModel.undo();
    ElementsTestRunner.waitForStyles('element', step3, true);
  }

  function step3() {
    TestRunner.addResult('=== After undo ===');
    ElementsTestRunner.dumpSelectedElementStyles(true);

    TestRunner.domModel.redo();
    ElementsTestRunner.waitForStyles('element', step4, true);
  }

  function step4() {
    TestRunner.addResult('=== After redo ===');
    ElementsTestRunner.dumpSelectedElementStyles(true);
    var section = UI.panels.elements._stylesWidget._sectionBlocks[1].sections[1];
    section.startEditingSelector();
    section._selectorElement.textContent = '1% /*';
    section._selectorElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
    ElementsTestRunner.waitForSelectorCommitted(step5);
  }

  function step5() {
    TestRunner.addResult('=== After invalid key modification ===');
    ElementsTestRunner.dumpSelectedElementStyles(true);
    TestRunner.completeTest();
  }
}
</script>
</head>

<body onload="runTest()">
<div id="element"></div>
<p>
Tests that keyframes are shown in styles pane.
</p>

</body>
</html>
